Base tutorial finished

+++ server/controllers/tutorial.js
@@ -13,7 +13,7 @@ export const getTutorial = async (req,res)=>{
 export const createPost = async (req,res)=>{
     const post = req.body;
 
-    const newPost = new PostMessage(post);
+    const newPost = new PostMessage({...post, creator: req.userId, createdAt: new Date().toISOString()});
 
     try{
         await newPost.save();
@@ -36,6 +36,9 @@ export const updatePost = async(req,res)=>{
 
 export const deletePost = async(req,res) =>{
     const {id} = req.params;
+
+    if(!req.userId) return res.json({message:'Unauthenticated'});
+
     if(!mongoose.Types.ObjectId.isValid(id)) return res.status(404).send('No post with that id');
     await PostMessage.findByIdAndRemove(id);
     res.json({message: 'Post deleted succesfully'});
